+++ CHANGELOG.md
@@ -4,8 +4,15 @@
 
 * Fixes
   * Security fix for [Misuse of public apis can result in private key exposure #13](https://github.com/potatosalad/erlang-libdecaf/issues/13) (see [report here](https://github.com/MystenLabs/ed25519-unsafe-libs)).
+* Library Support
+  * Legacy KECCAK SHA-3 (thanks to [@ukazap](https://github.com/ukazap))
+    * `libdecaf_keccak_sha3:hash/2`, `libdecaf_keccak_sha3:hash/3`
+    * `libdecaf_keccak_sha3:init/1`
+    * `libdecaf_keccak_sha3:update/2`
+    * `libdecaf_keccak_sha3:final/1`, `libdecaf_keccak_sha3:final/2`
 * Enhancements
   * New Keypair API for Ed25519 and Ed448 operations (see [#13](https://github.com/potatosalad/erlang-libdecaf/issues/13)).
+  * [Add legacy Keccak support #15](https://github.com/potatosalad/erlang-libdecaf/pull/15) (see also [#12](https://github.com/potatosalad/erlang-libdecaf/issues/12).)
   * Upstream update to [`ed448goldilocks` version `features-20220828`](https://github.com/potatosalad/ed448goldilocks/tree/features-20220828) (vendored as part of `libdecaf` app).
 
 ## 2.0.0 (2022-01-25)